Partial implementation of database upgrade and downgrade.
On downgrade, clear the database to allow creation of old schema. I don't think
we can do better since we cannot know the future schema.

On upgrade, we should copy old data to the new schema, but currently we simply
drop everything and create the database from scratch to make development easier.